task-13 SQL database tests and Jpa database tests not working (#22)

+++ app/src/main/java/pl/futurecollars/invoicing/db/file/IdService.java
@@ -8,7 +8,7 @@ public class IdService {
   private final Path idFilePath;
   private final FileService fileService;
 
-  private Integer nextId = 1;
+  private long nextId = 1;
 
   public IdService(Path idFilePath, FileService fileService) {
     this.idFilePath = idFilePath;
@@ -26,7 +26,7 @@ public class IdService {
 
   }
 
-  public int getNextIdAndIncrement() {
+  public Long getNextIdAndIncrement() {
     try {
       fileService.writeToFile(idFilePath, String.valueOf(nextId + 1));
       return nextId++;
